src/main/java/hello/jdbc/service/MemberServiceV3_3.java

package hello.jdbc.service;

import hello.jdbc.domain.Member;
import hello.jdbc.repository.MemberRepositroyV3;
import lombok.extern.slf4j.Slf4j;
import org.springframework.transaction.annotation.Transactional;

import java.sql.SQLException;

/**
 * 트랜잭션 - @Transactional AOP
 */
@Slf4j
public class MemberServiceV3_3 {

    private final MemberRepositroyV3 memberRepositroy;

    public MemberServiceV3_3(MemberRepositroyV3 memberRepositroy) {
        this.memberRepositroy = memberRepositroy;
    }

    @Transactional
    public void accountTransfer(String fromId, String toId, int money) throws SQLException {
        bizLogic(fromId, toId, money);
    }

    private void bizLogic(String fromId, String toId, int money) throws SQLException {
        Member fromMember = memberRepositroy.findById(fromId);
        Member toMember = memberRepositroy.findById(toId);

        memberRepositroy.update(fromId, fromMember.getMoney() - money);

        validation (toMember);

        memberRepositroy.update(toId, fromMember.getMoney() + money);
    }

    private void validation(Member toMember) {
        if (toMember.getMemberId().equals("ex")) {
            throw new IllegalStateException("이체중 예외 발생");
        }
    }
}
